move UnitSearchComponent into _shared module
