+++ src/app/report/report-parameter/_shared/unit-search/profiles-table/profiles-table.component.scss
